Add : UI
화면 상단에 출력될 타이머 기능 구현

+++ Assets/01_Scripts/05_UI/Managers/DataManager.cs
@@ -78,7 +78,7 @@ public class DataManager : MonoBehaviour
                 {
                     if (highScores.list[i].highScore < highScores.list[j].highScore)
                     {
-                        int temp = highScores.list[j].highScore;
+                        float temp = highScores.list[j].highScore;
                         highScores.list[j].highScore = highScores.list[i].highScore;
                         highScores.list[i].highScore = temp;
                     }
